Rename FileSerializationSink to just SerializationSink, now that there is a single kind.
