+++ analyzeme/benches/serialization_bench.rs
@@ -5,15 +5,15 @@ extern crate test;
 use analyzeme::testing_common;
 
 #[bench]
-fn bench_file_serialization_sink(bencher: &mut test::Bencher) {
+fn bench_serialization_sink(bencher: &mut test::Bencher) {
     bencher.iter(|| {
-        testing_common::run_serialization_bench("file_serialization_sink_test", 500_000, 1);
+        testing_common::run_serialization_bench("serialization_sink_test", 500_000, 1);
     });
 }
 
 #[bench]
-fn bench_file_serialization_sink_8_threads(bencher: &mut test::Bencher) {
+fn bench_serialization_sink_8_threads(bencher: &mut test::Bencher) {
     bencher.iter(|| {
-        testing_common::run_serialization_bench("file_serialization_sink_test", 50_000, 8);
+        testing_common::run_serialization_bench("serialization_sink_test_8_threads", 50_000, 8);
     });
 }
