+++ measureme/src/file_header.rs
@@ -1,7 +1,7 @@
 //! All binary files generated by measureme have a simple file header that
 //! consists of a 4 byte file magic string and a 4 byte little-endian version
 //! number.
-use crate::FileSerializationSink;
+use crate::SerializationSink;
 use std::convert::TryInto;
 use std::error::Error;
 
@@ -14,7 +14,7 @@ pub const FILE_MAGIC_STRINGTABLE_INDEX: &[u8; 4] = b"MMSI";
 /// rely on this size to be `8`.
 pub const FILE_HEADER_SIZE: usize = 8;
 
-pub fn write_file_header(s: &FileSerializationSink, file_magic: &[u8; 4]) {
+pub fn write_file_header(s: &SerializationSink, file_magic: &[u8; 4]) {
     // The implementation here relies on FILE_HEADER_SIZE to have the value 8.
     // Let's make sure this assumption cannot be violated without being noticed.
     assert_eq!(FILE_HEADER_SIZE, 8);
@@ -53,11 +53,11 @@ pub fn strip_file_header(data: &[u8]) -> &[u8] {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::file_serialization_sink::FileSerializationSink;
+    use crate::SerializationSink;
 
     #[test]
     fn roundtrip() {
-        let data_sink = FileSerializationSink::new_in_memory();
+        let data_sink = SerializationSink::new_in_memory();
 
         write_file_header(&data_sink, FILE_MAGIC_EVENT_STREAM);
 
@@ -71,7 +71,7 @@ mod tests {
 
     #[test]
     fn invalid_magic() {
-        let data_sink = FileSerializationSink::new_in_memory();
+        let data_sink = SerializationSink::new_in_memory();
         write_file_header(&data_sink, FILE_MAGIC_STRINGTABLE_DATA);
         let mut data = data_sink.into_bytes();
 
@@ -82,7 +82,7 @@ mod tests {
 
     #[test]
     fn other_version() {
-        let data_sink = FileSerializationSink::new_in_memory();
+        let data_sink = SerializationSink::new_in_memory();
 
         write_file_header(&data_sink, FILE_MAGIC_STRINGTABLE_INDEX);
